ssc build: fix race condition crash
Process handle is created in Process::open()
Process status is then checked in a thread
By the time this check happens, Process::open() has completed, resulting in an invalid handle
GetExitCodeProcess() would then return a spurious value of 4294967295, or sometimes other values.
stoi() would then crash with an out of range exception.

Solution:
1. Pass copy of thread handle to process status waiter thread
2. On GetExitCodeProcess() failure, call GetLastError() and write assosciated error message to stderr
The output will appear as follows:
```
• package prepared 683ms
Error 6 in SSC::Process::open() GetExitCodeProcess(): The handle is invalid.

• ran user build command 19ms
• build failed, exiting with code 4294967295 0ms
```

3. Use std::stoul() instead of std::stoi() in case any large return codes occur in other scenarios.

+++ src/cli/cli.cc
@@ -1778,10 +1778,10 @@ int main (const int argc, const char* argv[]) {
         [](SSC::String const &out) { stdWrite(out, false); },
         [](SSC::String const &out) { stdWrite(out, true); },
         [](SSC::String const &code) {
-          if (std::stoi(code) != 0) {
+          if (std::stoul(code) != 0) {
             log("build failed, exiting with code " + code);
             // TODO(trevnorris): Force non-windows to exit the process.
-            exit(std::stoi(code));
+            exit(std::stoul(code));
           }
         }
       );
